+++ common/src/main/java/com/avp/common/registry/init/creative_mode_tab/initializer/SpawnEggsCreativeModeTabInitializer.java
@@ -19,6 +19,7 @@ public class SpawnEggsCreativeModeTabInitializer {
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.FACEHUGGER_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.OVOMORPH_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.PRAETORIAN_SPAWN_EGG);
+        CreativeModeTabUtil.accept(output, AlienSpawnEggItems.PREDALIEN_ADOLESCENT_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.PREDALIEN_CHESTBURSTER_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.PREDALIEN_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.PROWLER_SPAWN_EGG);
@@ -35,6 +36,7 @@ public class SpawnEggsCreativeModeTabInitializer {
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.NETHER_FACEHUGGER_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.NETHER_OVOMORPH_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.NETHER_PRAETORIAN_SPAWN_EGG);
+        CreativeModeTabUtil.accept(output, AlienSpawnEggItems.NETHER_PREDALIEN_ADOLESCENT_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.NETHER_PREDALIEN_CHESTBURSTER_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.NETHER_PREDALIEN_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.NETHER_PROWLER_SPAWN_EGG);
@@ -51,6 +53,7 @@ public class SpawnEggsCreativeModeTabInitializer {
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.ABERRANT_FACEHUGGER_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.ABERRANT_OVOMORPH_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.ABERRANT_PRAETORIAN_SPAWN_EGG);
+        CreativeModeTabUtil.accept(output, AlienSpawnEggItems.ABERRANT_PREDALIEN_ADOLESCENT_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.ABERRANT_PREDALIEN_CHESTBURSTER_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.ABERRANT_PREDALIEN_SPAWN_EGG);
         CreativeModeTabUtil.accept(output, AlienSpawnEggItems.ABERRANT_PROWLER_SPAWN_EGG);